TUT-238 [BUGFIX] Add confirmation box

+++ app/client/containers/EndpointDoc/EndpointDoc.js
@@ -144,7 +144,9 @@ class EndpointDoc extends React.Component {
   }
 
   onRemoveUrlParam = (id) => {
-    this.props.dispatch(removeUrlParams(id));
+    if (confirm('are you sure')) {
+      this.props.dispatch(removeUrlParams(id));
+    }
   }
 
   renderResponses() {
